Rearrange the entire api so that size is a calculated thing. (#3)
When you think about it the size of a chunk can be calculated but it does not really
matter after we have read the data in. I have removed it from our representation so that
we will always have to calculate it. It leaves even less room for mistakes going forwards.